apps/api/requirements-dev.txt
Django>=4.2
djangorestframework
psycopg[binary]
drf-spectacular
django-environ
msal
requests
azure-identity
azure-storage-blob
azure-keyvault-secrets
django-cors-headers
pytest
pytest-django
pytest-cov
ruff
black
coverage
